@@ -3,33 +3,33 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "actioncable": "^5.2.4",
+    "actioncable": "^5.2.6",
     "foundation-sites": "6.6.3",
-    "grapnel": "^0.7.2",
-    "http-proxy-middleware": "^0.21.0",
+    "grapnel": "0.7.2",
+    "http-proxy-middleware": "^2.0.0",
     "material-ui": "^0.20.2",
-    "mixpanel-browser": "^2.39.0",
-    "node-sass": "^4.14.1",
+    "mixpanel-browser": "^2.41.0",
+    "node-sass": "^5.0.0",
     "prop-types": "^15.6.2",
-    "react": "16.14.0",
-    "react-autosize-textarea": "^6.0.0",
-    "react-dom": "16.14.0",
-    "react-google-login": "^4.0.1",
-    "react-helmet": "^5.2.0",
-    "react-hotkeys": "^1.1.4",
-    "react-redux": "^7.2.1",
-    "react-scripts": "^3.4.3",
-    "react-scroll": "^1.8.0",
-    "redux": "^4.0.5"
+    "react": "^17.0.2",
+    "react-autosize-textarea": "^7.1.0",
+    "react-dom": "^17.0.2",
+    "react-google-login": "^5.2.2",
+    "react-helmet": "^6.1.0",
+    "react-hotkeys": "^2.0.0",
+    "react-redux": "^7.2.4",
+    "react-scripts": "^4.0.3",
+    "react-scroll": "^1.8.2",
+    "redux": "^4.1.0"
   },
   "devDependencies": {
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint-config-airbnb": "^17.1.0",
-    "fast-deep-equal": "^2.0.1",
+    "eslint-config-airbnb": "^18.2.1",
+    "fast-deep-equal": "^3.1.3",
     "jest-enzyme": "^7.1.2",
     "promise-mock": "^2.1.0",
-    "serve": "^11.3.0"
+    "serve": "^11.3.2"
   },
   "scripts": {
     "start": "REACT_APP_USE_MOCK_GOOGLE=\"$USE_MOCK_GOOGLE\" REACT_APP_GOOGLE_OAUTH_CLIENT_ID=\"$GOOGLE_OAUTH_CLIENT_ID\" react-scripts start",
@@ -43,7 +43,7 @@
     "extends": "react-app"
   },
   "engines": {
-    "node": "^12.6.0"
+    "node": "^14.16.1"
   },
   "browserslist": [
     ">0.2%",
