@@ -31,6 +31,6 @@
 require 'clients/google_client'
 
 auth_endpoint = ENV['GOOGLE_AUTH_ENDPOINT'] ||= 'https://www.googleapis.com/oauth2/v3/userinfo'
-hosted_domain = ENV.fetch('HOSTED_DOMAIN') { nil }
+hosted_domain = ENV.fetch('HOSTED_DOMAIN', nil)
 
 GOOGLE_CLIENT = GoogleClient.new(auth_endpoint, hosted_domain)
