@@ -86,6 +86,17 @@ module SpecHelpers
     page.driver.execute_script("window.localStorage.setItem('authToken', 'secret-test-user-token-ef87c521b979');")
   end
 
+  def create_sharable_retro_as_admin(slug, password)
+    login_as_admin
+    click_on 'Retros'
+    click_on 'New Retro'
+    fill_in 'Name', with: 'Sharable retro'
+    fill_in 'Slug', with: slug
+    fill_in 'Password', with: password
+    page.check 'Share this retro using a magic link?'
+    click_on 'Create Retro'
+  end
+
   def create_private_retro(team_name = nil)
     fill_in_create_retro_form(team_name)
     find('#retro_is_magic_link_enabled', visible: :all).set(true)
@@ -181,12 +192,13 @@ module SpecHelpers
   end
 
   def mark_all_retro_items_as_done
-    sleep(0.3)
+    sleep(1)
     find_all('div.retro-item').each do |retro_item|
       item_id = retro_item[:id]
       non_flaky_click("##{item_id}")
-      sleep(0.3)
+      sleep(0.5)
       non_flaky_click("##{item_id} .item-done")
+      sleep(0.5)
     end
   end
 
