@@ -4,6 +4,7 @@ module.exports = {
     'airbnb',
   ],
   'rules': {
+    'arrow-parens': ['error', 'always'],
     'camelcase': ['off'],
     'class-methods-use-this': ['off'],
     'jsx-a11y/click-events-have-key-events': ['off'],
@@ -11,17 +12,20 @@ module.exports = {
     'jsx-a11y/label-has-for': ['off'],
     'jsx-a11y/no-autofocus': ['off'],
     'jsx-a11y/no-static-element-interactions': ['off'],
+    'max-classes-per-file': ['off'],
     'max-len': ['off'],
+    'no-multiple-empty-lines': ['off'],
+    'object-curly-newline': ['error', {
+      'consistent': true,
+    }],
+    'object-curly-spacing': ['error', 'never'],
+    'operator-linebreak': ['error', 'after'],
     'prefer-destructuring': ['off'],
+    'prefer-object-spread': ['off'],
     'prefer-template': ['off'],
+    'quote-props': ['error', 'consistent'],
     'react/destructuring-assignment': ['off'],
     'react/forbid-prop-types': ['off'],
-    'react/jsx-no-target-blank': ['off'],
-    'react/jsx-one-expression-per-line': ['off'],
-    'react/no-unescaped-entities': ['off'],
-    'react/sort-comp': ['off'],
-
-    'arrow-parens': ['error', 'always'],
     'react/jsx-no-bind': ['error', {
       'ignoreDOMComponents': false,
       'ignoreRefs': false,
@@ -29,18 +33,18 @@ module.exports = {
       'allowFunctions': false,
       'allowBind': false,
     }],
-    'object-curly-newline': ['error', {
-      'consistent': true,
-    }],
-    'object-curly-spacing': ['error', 'never'],
-    'operator-linebreak': ['error', 'after'],
+    'react/jsx-no-target-blank': ['off'],
+    'react/jsx-one-expression-per-line': ['off'],
+    'react/jsx-props-no-spreading': ['off'],
     'react/jsx-tag-spacing': ['error', {
       'closingSlash': 'never',
       'beforeSelfClosing': 'never',
       'afterOpening': 'never',
       'beforeClosing': 'never',
     }],
-    'quote-props': ['error', 'consistent'],
+    'react/no-unescaped-entities': ['off'],
+    'react/sort-comp': ['off'],
+    'react/static-property-placement': ['off'],
   },
 
   'overrides': [{
