@@ -197,7 +197,7 @@ describe('inputting an action item', () => {
     subject.find('.emoji-button').simulate('click', {});
     subject.find('.emoji-selector-option').first().simulate('mouseDown', {});
 
-    expect(textarea).toHaveText(String.fromCodePoint(0x1F600) + 'anything');
+    expect(textarea).toHaveText(`anything${String.fromCodePoint(0x1F600)}`);
   });
 
   it('hides the emoji selector on blur', () => {
