@@ -1,4 +1,4 @@
-const proxy = require('http-proxy-middleware');
+const {createProxyMiddleware: proxy} = require('http-proxy-middleware');
 
 module.exports = (app) => {
   app.use(proxy('/cable', {target: 'http://localhost:4000', ws: true}));
